+++ Packages/best.crux.core/CHANGELOG.md
@@ -7,6 +7,13 @@ A changelog wasn't kept until version 0.7.0, so the changelog is currently being
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- DecoratedList<T> can now return List<T>.Enumerator from its GetEnumerator method
+  - This is very nitpicky, but this prevents some unnecessary memory allocation
+
 ## [0.8.1] - 2025-10-14
 
 ### Added